+++ activerecord/lib/active_record/migration.rb
@@ -1185,10 +1185,10 @@ module ActiveRecord
       def run_without_lock
         migration = migrations.detect { |m| m.version == @target_version }
         raise UnknownMigrationVersionError.new(@target_version) if migration.nil?
-        result = execute_migration_in_transaction(migration, @direction)
 
         record_environment
-        result
+
+        execute_migration_in_transaction(migration, @direction)
       end
 
       # Used for running multiple migrations up to or down to a certain value.
@@ -1197,12 +1197,11 @@ module ActiveRecord
           raise UnknownMigrationVersionError.new(@target_version)
         end
 
-        result = runnable.each do |migration|
+        record_environment
+
+        runnable.each do |migration|
           execute_migration_in_transaction(migration, @direction)
         end
-
-        record_environment
-        result
       end
 
       # Stores the current environment in the database.