+++ activerecord/lib/active_record/connection_adapters/abstract_adapter.rb
@@ -74,7 +74,7 @@ module ActiveRecord
       SIMPLE_INT = /\A\d+\z/
 
       attr_accessor :visitor, :pool
-      attr_reader :schema_cache, :owner, :logger
+      attr_reader :schema_cache, :owner, :logger, :prepared_statements, :lock
       alias :in_use? :owner
 
       def self.type_cast_config_to_integer(config)
@@ -92,8 +92,6 @@ module ActiveRecord
           config
         end
       end
-
-      attr_reader :prepared_statements
 
       def initialize(connection, logger = nil, config = {}) # :nodoc:
         super()