+++ activerecord/lib/active_record/connection_adapters/mysql/schema_statements.rb
@@ -38,6 +38,7 @@ module ActiveRecord
 
               indexes.last.columns << row[:Column_name]
               indexes.last.lengths.merge!(row[:Column_name] => row[:Sub_part].to_i) if row[:Sub_part]
+              indexes.last.orders.merge!(row[:Column_name] => :desc) if row[:Collation] == "D"
             end
           end