+++ activerecord/lib/active_record/migration.rb
@@ -1105,6 +1105,10 @@ module ActiveRecord
       def move(direction, migrations_paths, steps)
         migrator = new(direction, migrations(migrations_paths))
 
+        if current_version != 0 && !migrator.current_migration
+          raise UnknownMigrationVersionError.new(current_version)
+        end
+
         start_index =
           if current_version == 0
             0
@@ -1112,11 +1116,9 @@ module ActiveRecord
             migrator.migrations.index(migrator.current_migration)
           end
 
-        if start_index
-          finish = migrator.migrations[start_index + steps]
-          version = finish ? finish.version : 0
-          send(direction, migrations_paths, version)
-        end
+        finish = migrator.migrations[start_index + steps]
+        version = finish ? finish.version : 0
+        send(direction, migrations_paths, version)
       end
     end