+++ activerecord/CHANGELOG.md
@@ -1,3 +1,13 @@
+*   Record environment in database before running migrations
+
+    The environment is now set in the database before running DB migrations,
+    so in the event of a migration error on a new database, destructive rake
+    tasks can be run.
+
+    Fixes #28001.
+
+    *Dominic Cleal*
+
 *   Load only needed records on `ActiveRecord::Relation#inspect`.
 
     Instead of loading all records and returning only a subset of those, just