+++ activerecord/CHANGELOG.md
@@ -1,3 +1,11 @@
+*   When multiple threads are sharing a database connection inside a test using
+    transactional fixtures, a nested transaction will temporarily lock the
+    connection to the current thread, forcing others to wait.
+
+    Fixes #28197.
+
+    *Matthew Draper*
+
 *   Load only needed records on `ActiveRecord::Relation#inspect`.
 
     Instead of loading all records and returning only a subset of those, just