+++ activerecord/CHANGELOG.md
@@ -1,1 +1,5 @@
+*   Quote database name in db:create grant statement (when database_user does not have access to create the database).
+
+    *Rune Philosof*
+
 Please check [5-1-stable](https://github.com/rails/rails/blob/5-1-stable/activerecord/CHANGELOG.md) for previous changes.