+++ activerecord/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Quote database name in db:create grant statement (when database_user does not have access to create the database).
+
+    *Rune Philosof*
+
 *   Raise error `UnknownMigrationVersionError` on the movement of migrations
     when the current migration does not exist.